FEATURE: moves getDifficulty
