+++ libs/spirit-islander/util/src/lib/game/get-difficulty.ts
@@ -1,5 +1,5 @@
-import type { Difficulty } from '../game/difficulty';
-import type { ExpansionName } from '../game/expansions';
+import type { Difficulty } from './difficulty';
+import type { ExpansionName } from './expansions';
 
 export function getDifficulty(
   difficulty: Difficulty | ((expansions: ExpansionName[]) => Difficulty),
